+++ supvisors/test/use_cases/scenario_1/bin/scen1/sensor_acquisition.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+test_dir=$(dirname "$(readlink -f "$0")")
+
+$test_dir/common.sh ${SUPERVISOR_PROCESS_NAME}