+++ supvisors/test/use_cases/scenario_3/bin/common_bus_interface.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-./bin/common.sh ${SUPERVISOR_PROCESS_NAME}